- remote execution

+++ OSK/OrleansTestApplication/Program.cs
@@ -10,7 +10,7 @@ namespace OrleansTestApplication
 {
     class Program
     {
-        public static int Main(string[] args) =>  (new TestRemoteExecution()).RunMainAsync().Result;
+        public static int Main(string[] args) =>  (int)(new TestRemoteExecution()).RunMainAsync().Result;
 
 
     }
